Moved GlobalExceptionHandler to error package, removed the Auth token
