+++ Project/app/src/main/java/org/ionproject/android/common/IonApplication.kt
@@ -12,6 +12,7 @@ import org.ionproject.android.common.ionwebapi.JacksonIonMapper
 import org.ionproject.android.common.repositories.*
 import org.ionproject.android.common.workers.WorkerManagerFacade
 import org.ionproject.android.error.ErrorActivity
+import org.ionproject.android.error.GlobalExceptionHandler
 import retrofit2.Retrofit
 import retrofit2.converter.scalars.ScalarsConverterFactory
 
@@ -45,10 +46,11 @@ class IonApplication : Application() {
 
         crashlytics = FirebaseCrashlytics.getInstance()
 
-        globalExceptionHandler = GlobalExceptionHandler { thread, throwable ->
-            val intent = Intent(applicationContext, ErrorActivity::class.java)
-            applicationContext.startActivity(intent)
-        }
+        globalExceptionHandler =
+            GlobalExceptionHandler { thread, throwable ->
+                val intent = Intent(applicationContext, ErrorActivity::class.java)
+                applicationContext.startActivity(intent)
+            }
 
         /**
          * Our app runs in a single process therefore we follow
