+++ Project/app/src/main/java/org/ionproject/android/common/ionwebapi/IonService.kt
@@ -4,7 +4,7 @@ import retrofit2.http.GET
 import retrofit2.http.Headers
 import retrofit2.http.Url
 
-private const val AUTHORIZATION_TOKEN = "5tuMrTm8dD-n2d53e3z2ALYC5JRUshAhZUa7xp3cHH4"
+private const val AUTHORIZATION_TOKEN = "ASK_THE_DEVELOPERS"
 
 interface IonService {
     @Headers("Authorization: Bearer $AUTHORIZATION_TOKEN")
